# app/backend/src/api-documentation/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Traveling Merchant",
    "version": "1",
    "description": "A back end application to keep track of transactions of a merchant buying product and selling product and keeping track of their inventory. All succesful operations start with \"Merchant: {merchantId}\" where merchantId is the id of the merchant that is logged in.",
    "license": {
      "name": "MIT",
      "url": "https://spdx.org/licenses/MIT.html"
    }
  },
  "servers": [
    {
      "url": "http://localhost:{PORT}",
      "variables": {},
      "description": "The application's URL is randomized after every initialization through its port number represented in the URL as {PORT}."
    }
  ],
  "paths": {
...
    "/orders/{orderId}": {
      "summary": "Request for a specific order's data",
      "description": "This path used to perform CRUD operations on orders",
      "parameters": [
        {
          "in": "path",
          "name": "orderId",
          "required": true,
          "schema": {
            "type": "integer",
            "minimum": 1
          },
          "description": "The order ID"
        }
      ],
      "get": {
        "summary": "Get order",
        "description": "This path uses the GET operation to get specific orders.",
        "operationId": "",
        "responses": {
          "200": {
            "description": "OK, returns order data with associated provider, wares bought, and expenses.",
            "content": {
              "application/json": {
                "examples": {
                  "Order found given order id": {
                    "value": "{\"id\":2,\"providerId\":2,\"cost\":959.59,\"tax\":89.59,\"shipment\":20,\"expectedAt\":\"2025-01-09T00:00:00.000Z\",\"actualAt\":\"2025-01-09T00:00:00.000Z\",\"createdAt\":\"2025-01-01T00:00:00.000Z\",\"updatedAt\":\"2025-01-17T00:00:00.000Z\",\"provider\":{\"id\":2,\"name\":\"Ebay\",\"address\":\"0000 online\",\"phoneNumber\":\"5125869601\",\"email\":\"\",\"createdAt\":\"2025-01-01T00:00:00.000Z\",\"updatedAt\":\"2025-01-01T00:00:00.000Z\"},\"ordersWares\":[{\"id\":4,\"wareId\":4,\"orderId\":2,\"unitPrice\":150,\"amount\":5,\"returned\":0,\"createdAt\":\"2025-01-01T00:00:00.000Z\",\"updatedAt\":\"2025-01-01T00:00:00.000Z\",\"ware\":{\"id\":4,\"name\":\"Versace Men's 4-Pc. Eros Eau de Toilette Gift Set\",\"type\":\"perfume\",\"tags\":[\"men\",\"4-pc\"],\"unitPrice\":176,\"createdAt\":\"2025-01-09T00:00:00.000Z\",\"updatedAt\":\"2025-01-09T00:00:00.000Z\",\"stock\":5}},{\"id\":5,\"wareId\":5,\"orderId\":2,\"unitPrice\":10,\"amount\":10,\"returned\":0,\"createdAt\":\"2025-01-01T00:00:00.000Z\",\"updatedAt\":\"2025-01-01T00:00:00.000Z\",\"ware\":{\"id\":5,\"name\":\"Eymi Unisex Leather Braclet with Infinity Sign Symbolic Love Fashion Braided Wristband Bangle\",\"type\":\"braclet\",\"tags\":[\"unisex\"],\"unitPrice\":14,\"createdAt\":\"2025-01-09T00:00:00.000Z\",\"updatedAt\":\"2025-01-09T00:00:00.000Z\",\"stock\":8}}]}"
                  }
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND",
            "content": {
              "application/json": {
                "examples": {
                  "Order not found given order id": {
                    "value": "Order not found."
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Orders"
        ]
      },
      "put": {
        "summary": "Update order",
        "description": "This path uses PUT operation to update orders information.",
        "requestBody": {
          "description": "Order new information. The date properties (has the property name end with \"At\") should be in the following formats 2024-12, 2024-12-02, or \"2024-12-02T04:32:42.456Z\".",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ordersUpdates"
              }
            }
          }
        },
        "operationId": "",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "examples": {
                  "Updates one piece of information": {
                    "value": "Merchant: 1 order with id = 1 was updated with given shipment = 577."
                  },
                  "Updates multiple pieces of order information": {
                    "value": "Merchant: 1 order with id = 4 was updated with given providerId = 9, cost = 733, and tax = 631."
                  },
                  "Updates all of an exisitng order information": {
                    "value": "Merchant: 1 order with id = 2 was updated with given providerId = 9, cost = 556, tax = 745, shipment = 619, expectedAt = Mon Sep 08 2025 15:44:47 GMT-0500 (Central Daylight Time), and actualAt = Sun Oct 12 2025 12:34:40 GMT-0500 (Central Daylight Time)."
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST",
            "content": {
              "application/json": {
                "examples": {
                  "No inputs are given in the request body": {
                    "value": "Bad input request."
                  }
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND",
            "content": {
              "application/json": {
                "examples": {
                  "Order not found given order id": {
                    "value": "Order not found."
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Orders"
        ]
      },
      "delete": {
        "summary": "Delete order",
        "description": "This path uses the DELETE operation to delete orders from database.",
        "operationId": "",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "examples": {
                  "Order is succesfully deleted": {
                    "value": "Merchat: 1 has deleted a order with id = 3."
                  }
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND",
            "content": {
              "application/json": {
                "examples": {
                  "Order not found given order id": {
                    "value": "Order not found."
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ordersUpdates": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1
          },
          "providerId": {
            "type": "integer",
            "minimum": 1
          },
          "cost": {
            "type": "number",
            "minimum": 0
          },
          "tax": {
            "type": "number",
            "minimum": 0
          },
          "shipment": {
            "type": "number",
            "minimum": 0
          },
          "expectedAt": {
            "type": "string"
          },
          "actualAt": {
            "type": "string"
          }
        },
        "xml": {
          "name": "Updated order information"
        }
      },
...
    }
  }
}